codebackup/test_app/260420 로그인코드완료/requirements.txt
Appium-Python-Client==5.3.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
opencv-python==4.13.0.92
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
PySocks==1.7.1
pytest==8.1.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
python-dotenv==1.2.2
requests==2.31.0
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
